+++ packages/bridge-controller/src/utils/fetch.test.ts
@@ -81,6 +81,7 @@ describe('fetch', () => {
         BridgeClientId.EXTENSION,
         mockFetchFn,
         BRIDGE_PROD_API_BASE_URL,
+        '1.0.0',
       );
 
       expect(mockFetchFn).toHaveBeenCalledWith(
@@ -90,7 +91,7 @@ describe('fetch', () => {
             cacheRefreshTime: 600000,
           },
           functionName: 'fetchBridgeTokens',
-          headers: { 'X-Client-Id': 'extension' },
+          headers: { 'X-Client-Id': 'extension', 'Client-Version': '1.0.0' },
         },
       );
 
@@ -145,6 +146,7 @@ describe('fetch', () => {
           BridgeClientId.EXTENSION,
           mockFetchFn,
           BRIDGE_PROD_API_BASE_URL,
+          '1.0.0',
         ),
       ).rejects.toThrow(mockError);
     });
@@ -175,6 +177,7 @@ describe('fetch', () => {
         mockFetchFn,
         BRIDGE_PROD_API_BASE_URL,
         null,
+        '1.0.0',
       );
 
       expect(mockFetchFn).toHaveBeenCalledWith(
@@ -184,7 +187,7 @@ describe('fetch', () => {
             cacheRefreshTime: 0,
           },
           functionName: 'fetchBridgeQuotes',
-          headers: { 'X-Client-Id': 'extension' },
+          headers: { 'X-Client-Id': 'extension', 'Client-Version': '1.0.0' },
           signal,
         },
       );
@@ -197,6 +200,7 @@ describe('fetch', () => {
       );
       expect(result.validationFailures).toStrictEqual([]);
       expect(mockConsoleWarn).not.toHaveBeenCalled();
+      mockConsoleWarn.mockRestore();
     });
 
     it('should fetch bridge quotes successfully, with approvals', async () => {
@@ -235,6 +239,7 @@ describe('fetch', () => {
         mockFetchFn,
         BRIDGE_PROD_API_BASE_URL,
         null,
+        '1.0.0',
       );
 
       expect(mockFetchFn).toHaveBeenCalledWith(
@@ -244,7 +249,7 @@ describe('fetch', () => {
             cacheRefreshTime: 0,
           },
           functionName: 'fetchBridgeQuotes',
-          headers: { 'X-Client-Id': 'extension' },
+          headers: { 'X-Client-Id': 'extension', 'Client-Version': '1.0.0' },
           signal,
         },
       );
@@ -260,6 +265,7 @@ describe('fetch', () => {
         'socket|trade',
       ]);
       expect(mockConsoleWarn).toHaveBeenCalledTimes(1);
+      mockConsoleWarn.mockRestore();
     });
 
     it('should filter out malformed bridge quotes', async () => {
@@ -313,6 +319,7 @@ describe('fetch', () => {
         mockFetchFn,
         BRIDGE_PROD_API_BASE_URL,
         null,
+        '1.0.0',
       );
 
       expect(mockFetchFn).toHaveBeenCalledWith(
@@ -322,7 +329,7 @@ describe('fetch', () => {
             cacheRefreshTime: 0,
           },
           functionName: 'fetchBridgeQuotes',
-          headers: { 'X-Client-Id': 'extension' },
+          headers: { 'X-Client-Id': 'extension', 'Client-Version': '1.0.0' },
           signal,
         },
       );
@@ -360,6 +367,7 @@ describe('fetch', () => {
       `);
       // eslint-disable-next-line jest/no-restricted-matchers
       expect(mockConsoleWarn.mock.calls).toMatchSnapshot();
+      mockConsoleWarn.mockRestore();
     });
 
     it('should fetch bridge quotes successfully, with aggIds, bridgeIds and noFee=true', async () => {
@@ -386,6 +394,7 @@ describe('fetch', () => {
         mockFetchFn,
         BRIDGE_PROD_API_BASE_URL,
         FeatureId.PERPS,
+        '1.0.0',
       );
 
       expect(mockFetchFn).toHaveBeenCalledWith(
@@ -395,7 +404,7 @@ describe('fetch', () => {
             cacheRefreshTime: 0,
           },
           functionName: 'fetchBridgeQuotes',
-          headers: { 'X-Client-Id': 'extension' },
+          headers: { 'X-Client-Id': 'extension', 'Client-Version': '1.0.0' },
           signal,
         },
       );
@@ -435,6 +444,7 @@ describe('fetch', () => {
         baseUrl: 'https://api.example.com',
         fetchFn: mockFetchFn,
         clientId: 'test',
+        clientVersion: '1.0.0',
         assetIds: new Set([
           'eip155:1/erc20:0x123',
           'eip155:1/erc20:0x456',
@@ -459,7 +469,7 @@ describe('fetch', () => {
       expect(mockFetchFn).toHaveBeenCalledWith(
         'https://price.api.cx.metamask.io/v3/spot-prices?assetIds=eip155%3A1%2Ferc20%3A0x123%2Ceip155%3A1%2Ferc20%3A0x456&vsCurrency=USD',
         {
-          headers: { 'X-Client-Id': 'test' },
+          headers: { 'X-Client-Id': 'test', 'Client-Version': '1.0.0' },
           cacheOptions: { cacheRefreshTime: 30000 },
           functionName: 'fetchAssetExchangeRates',
         },
@@ -467,7 +477,7 @@ describe('fetch', () => {
       expect(mockFetchFn).toHaveBeenCalledWith(
         'https://price.api.cx.metamask.io/v3/spot-prices?assetIds=eip155%3A1%2Ferc20%3A0x123%2Ceip155%3A1%2Ferc20%3A0x456&vsCurrency=EUR',
         {
-          headers: { 'X-Client-Id': 'test' },
+          headers: { 'X-Client-Id': 'test', 'Client-Version': '1.0.0' },
           cacheOptions: { cacheRefreshTime: 30000 },
           functionName: 'fetchAssetExchangeRates',
         },
@@ -480,6 +490,7 @@ describe('fetch', () => {
         baseUrl: 'https://api.example.com',
         fetchFn: mockFetchFn,
         clientId: 'test',
+        clientVersion: '1.0.0',
         assetIds: new Set([
           'eip155:1/erc20:0x123',
           'eip155:1/erc20:0x456',
@@ -504,6 +515,7 @@ describe('fetch', () => {
         baseUrl: 'https://api.example.com',
         fetchFn: mockFetchFn,
         clientId: 'test',
+        clientVersion: '1.0.0',
         assetIds: new Set([
           'eip155:1/erc20:0x123',
           'eip155:1/erc20:0x456',
@@ -529,6 +541,7 @@ describe('fetch', () => {
         baseUrl: 'https://api.example.com',
         fetchFn: mockFetchFn,
         clientId: 'test',
+        clientVersion: '1.0.0',
         assetIds: new Set([
           'eip155:1/erc20:0x123',
           'eip155:1/erc20:0x456',
@@ -565,6 +578,7 @@ describe('fetch', () => {
         baseUrl: 'https://api.example.com',
         fetchFn: mockFetchFn,
         clientId: 'test',
+        clientVersion: '1.0.0',
         assetIds: new Set([
           'eip155:1/erc20:0x123',
           'eip155:1/erc20:0x456',
@@ -595,6 +609,7 @@ describe('fetch', () => {
         baseUrl: 'https://api.example.com',
         fetchFn: mockFetchFn,
         clientId: 'test',
+        clientVersion: '1.0.0',
         assetIds: new Set([
           'eip155:1/erc20:0x123',
           'eip155:1/erc20:0x456',
@@ -621,6 +636,7 @@ describe('fetch', () => {
         baseUrl: 'https://api.example.com',
         fetchFn: mockFetchFn,
         clientId: 'test',
+        clientVersion: '1.0.0',
         assetIds: new Set([
           'eip155:1/erc20:0x123',
           'eip155:1/erc20:0x456',
@@ -639,6 +655,7 @@ describe('fetch', () => {
         baseUrl: 'https://api.example.com',
         fetchFn: mockFetchFn,
         clientId: 'test',
+        clientVersion: '1.0.0',
         assetIds: new Set([]) as Set<CaipAssetType>,
       };
 
@@ -659,6 +676,7 @@ describe('fetch', () => {
         baseUrl: 'https://api.example.com',
         fetchFn: mockFetchFn,
         clientId: 'test',
+        clientVersion: '1.0.0',
         assetIds: new Set([
           'eip155:1/erc20:0x123',
           'eip155:1/erc20:0x456',